# swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
        "/v1/log/:level": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "level",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "String"
                        }
                    }
                ],
                "tags": [
                    "log"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LogIn"
                            }
                        }
                    },
                    "description": "LogIn",
                    "required": true
                }
            }
        },
        "/v1/log/project/:project": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "String"
                        }
                    }
                ],
                "tags": [
                    "log"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LogListIn"
                            }
                        }
                    },
                    "description": "LogListIn",
                    "required": true
                }
            }
        },
        "/v1/user/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "tags": [
                    "user"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginIn"
                            }
                        }
                    },
                    "description": "LoginIn",
                    "required": true
                }
            }
        },
        "/v1/user": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "tags": [
                    "user"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserIn"
                            }
                        }
                    },
                    "description": "UserIn",
                    "required": true
                }
            }
        },
        "/v1/user/current": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "tags": [
                    "user"
                ]
            }
        },
        "/v1/user/logout": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "tags": [
                    "user"
                ]
            }
        },
        "/v1/user/projects": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "tags": [
                    "user"
                ]
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "tags": [
                    "user"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Object"
                            }
                        }
                    },
                    "description": "Object",
                    "required": true
                }
            }
        },
        "/v1/user/forgot-password": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "tags": [
                    "user"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ForgotPasswordIn"
                            }
                        }
                    },
                    "description": "ForgotPasswordIn",
                    "required": true
                }
            }
        },
        "/v1/user/reset-password": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "tags": [
                    "user"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResetPasswordIn"
                            }
                        }
                    },
                    "description": "ResetPasswordIn",
                    "required": true
                }
            }
        },
        "/v1/project/projects": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "tags": [
                    "project"
                ]
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "tags": [
                    "project"
                ]
            }
        }
    },
    "openapi": "3.0.0",
    "components": {
        "schemas": {
            "LogIn": {
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "stackTrace": {
                        "type": "string"
                    }
                },
                "type": "object",
                "required": [
                    "message",
                    "stackTrace"
                ]
            },
            "LogListIn": {
                "properties": {
                    "page": {
                        "type": "string"
                    },
                    "size": {
                        "type": "string"
                    }
                },
                "type": "object",
                "required": [
                    "page",
                    "size"
                ]
            },
            "UserIn": {
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "type": "object",
                "required": [
                    "username",
                    "email",
                    "password"
                ]
            },
            "LoginIn": {
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "type": "object",
                "required": [
                    "username",
                    "password"
                ]
            },
            "ForgotPasswordIn": {
                "properties": {
                    "username": {
                        "type": "string"
                    }
                },
                "type": "object",
                "required": [
                    "username"
                ]
            },
            "ResetPasswordIn": {
                "properties": {
                    "token": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "type": "object",
                "required": [
                    "token",
                    "password"
                ]
            }
        }
    }
}